wk2/text_clustering_visualization_tsne/pyproject.toml
[project]
name = "text-clustering-visualization-tsne"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=3.6.0",
    "sentence-transformers>=4.1.0",
    "svlearn-bootcamp>=0.1.7",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/svlearn_text_clustering_visualization"]
